requirements.txt
gitdb==0.6.4
GitPython==1.0.1
smmap==0.9.0
wheel==0.24.0
